Move blockchainStore to database layer
